+Extension method to convert an ArgumentPosition enum to an index
We'll use this to turn a symbolic enum value into an index we can use
to get into an array. We'll use this to safely convert, instead of
sprinkling logic to cast and possibly subtract. This keeps us safe in
the event the enum values change to something less predictable.