Marked properties now accept the instance object that owns the property
Not sure what I was thinking about previously... It passed in the
attribute instance, which meant that the ATTRIBUTE property was being
set, not the argument class, which makes no sense. The whole point is
to match rules and poke values into the argument class.

This probably isn't its final form, but it gets closer to what we want.