+Spacing enum to describe space between the prefix and argument string
This is used in conjunction with the prefix string to describe either
a space-separated argument (-f filename), or having them joined into
one string (/f:filename). This lets the user be specific, and lets us
easily distinguish between the two.

This is also an enum--instead of a bool--so it's really clear what it
means they use it on the Argument attribute. It SAYS Spacing; it's not
an errant true or false running around with no clarity what it does.