+ClassBuilder blank class
This will be a builder-style object that can be used in a test to
construct a new TYPE of object, adding properties and attributes to it,
then finally catalyzing it into a real type. We'll then use this with
the ArgumentAnalyzer to match arguments.

The idea here is that we will dynamically generate the argument classes
instead of creating one real class (and file and project entry) for
every case we care about. Using this, we can define them per test and
then throw them away.