+Create factory method to initialize an instance as a builder
Users will go through this method (instead of a constructor) to obtain
an instance. This allows the class to do setup and pass any constructor
parameters it needs--in this case, a TypeBuilder.

It starts by defining the new dynamic assembly, the module, and the new
type (whose name is generated, since we don't care about the name). It
hands the TypeBuilder off to a new instance and kicks it back.

Next steps, the caller would call more things on this ClassBuilder to
add more things to the object we're building--properties, then
attributes, etc.