Can now attach attribute instances to properties dynamically
No error checking! :D

By how attributes work in reflection emit, it doesn't seem like we can
just pass an Attribute instance; instead we have to deal in the
CustomAttributeBuilder. So we use an expression and use that to
deconstruct the "new" call inside it to figure out what constructor is
being called and with what parameters. I don't think that extra
property sets will be covered by this, but this is a start. Usage looks
like this:

  AddAttribute( "FileName", () => new ArgumentAttribute( ... ) )