+ObjectExtensions class and Property extension method
This is designed to work with a fluent assertion, providing the value
of a property (by property name) using reflection.

For these dynamic argument classes, we can't ever hold a strongly-type
instance of it in code--we can only get at it through reflection. So as
the argument analyzer spits them out, we can't "dot" into it to assert
what its properties are. Instead, we'll use this to make a nicer syntax
even though it's straight reflection in the end.