+ArgumentAnalyzerHelper static class for testing
This exists to help facilitate using the ArgumentAnalyzer. That guy's
Analyze<T> method accepts a generic parameter, which must be known at
compile time, something that we don't have anymore with these dynamic
classes.

We'll only use this class in tests to facilitate calling that Analyze
method WITHOUT a generic parameter, even though we have the Type
instance that would satisfy it.