Adding meaningful error message around adding attributes too early
This is the case where you try to add an attribute to a property that
doesn't exist (hasn't been defined, misspelled, etc). It wouldn't find
the cached PropertyBuilder and would fail with a key not found. Now it
proactively checks for this and throws a more meaningful message to the
caller.