+ClassBuilder blank test class
Since the tests rely so heavily on this mechanism for testing the main
class in the entire project, it seems prudent to put it under test,
especially with all the complex moving parts it has.

The weird part is that ClassBuilder is a test-only class... and its
tests are in the same assembly as the production code tests. In a way,
it feels like it needs a "Test TEST" project. It may be valuable enough
to put in its own project, but for now, they're living together in
separate folders.