Making default constructor public--both were internal
Otherwise no one would be able to make an instance of this thing. Feels
a little weird to have an internal constructor just for testing purposes,
but I think it's weirder to expose an alternate path that isn't intended
to be used by the user. These classes weren't meant for third-party
extension, and it would only screw stuff up.