+GlobalCallbackTable static class
This represents a table of strings to Actions. This lets us stash callbacks
by name in a global repository that anyone can access. We'll use this when
creating dynamic methods to sub in function bodies.

Instead of munging the IL, we'll stash the supplied Action into this table
and have the dynamic method call into this global table.