+IMarkedMethod blank interface and blank concrete
This will be the method analog to the MarkedProperty--it represents a
method itself that has been decorated with the specified attribute type.
We'll use this when finding methods for specific attributes, and which
will let us get some info about it, and then invoke it.

This also serves as an abstraction for MethodInfo.