Adjusting TypeInspector to accept a Type, instead of Type argument
This is an internal class, so it's not like the user will have to deal
with this. Type parameters are nicer to use on the caller, instead of
calling GetType() or typeof(), but it makes it harder on dynamic stuff
where we don't actually know the type at compile time.

This should make testing a lot easier.