Can now call a handler with an error if present
This allows the user to add an error type parameter to their handler to
receive the reason things went wrong. This can be a good starting place
in case we expand this to be more like event args, instead of a single
enum value.

This should complete a few cases:

- No handler is found so none is invoked
- Multiple handlers are found and it rejects them
- One handler is found with no parameters and is called
- One handler is found with error parameter and is called

Note that this dupes a lot of the logic, so refactoring comes up next.