Moving ArgumentAnalyzerHelper into TestCommon
Unit and integration test projects used to have their own copy of this, since both projects needed it. The trouble was that it needed to ask ArgumentAnalyzer about its methods, which was NOT defined in TestCommon, so the assembly reference was an issue.

To "fix" this, the helper now takes in a type T that describes the ArgumentAnalyzer. So now the consumers pass in the type, instead of the helper directly knowing about the type.

This looks real dumb, but it's a way to reduce redundancy. It's also only test code, so I'm willing to live with this.